Split border variants into separate component files
Move each border style (Corners, Brackets, None) into its own file
under src/components/borders/. The top-level Border component now
just switches on BorderStyle and delegates to the variant components.
